feat(wizard): global-only settings + dedicated Projekt surface
Simplify the setup wizard's final step and split project-scoped config
out of the global flow:

- Review step: drop the "Install plan" file-count list and the
  per-file conflict-resolution UI (install.py governs overwrite; the
  list was informational only). Delete the now-dead WizardConflicts
  component + its test and the install-plan fetch/signals/props.
- Drop the "Where should these settings be written?" scope chooser.
  Settings and AI Council always write to the global tree (the
  ai-council write was already global); no project-scope branch on
  finish/apply.
- Move the project `modules` step out of the wizard into a new
  top-level "Projekt" surface (between Setup and Tasks): new
  ProjectSettingsPage + /project route, backed by a standalone
  POST /api/v1/modules/apply endpoint (and the ungated
  /api/v1/modules/detect). Writes .agent-project-settings.yml only.

Tests: replace the obsolete wizard modules-step test with
ProjectSettingsPage + modules-apply coverage. Builds on the in-flight
wizard-flow reorder (welcome → ai-tools → roles → packs lead).

Verified: vite build, tsc --noEmit, ESLint, full vitest (62 files /
417 tests) green.
